refactor: modularize agent logic by extracting SQL and visualization utilities into a dedicated tools package
