+++ src/ravioli/ai/agents/soul.md
